fix: concurrency safety, SD mutex balance, and worker build break
- nest_registry: defer logging out of the gLock critical section.
  findOrAddWorker now reports evict/drop via an out-param and callers
  log after taskEXIT_CRITICAL; Serial.printf with interrupts disabled
  could block on the UART lock and panic when the registry is full.
- nest_upload: track SD mutex ownership (sdHeld) so rejected uploads no
  longer sdGive() a mutex they never took.
- nest_upload: refuse worker syncs (raw 8080 + HTTP /upload) while a home
  upload holds the radio in STA mode, avoiding a WiFi teardown race.
- worker_config: cast Preferences::getInt() to int in max(); std::max on
  ESP32 core 3.x rejects mixed int/int32_t, which broke the worker build.

Both sketches verified compiling on esp32 core 3.3.8.

+++ firmware/worker/worker_config.cpp
@@ -123,8 +123,8 @@ void loadNvsConfig() {
   if (prefs.isKey("nestPsk"))  { prefs.getString("nestPsk",  nestPsk,  sizeof(nestPsk));  loaded++; }
   if (prefs.isKey("nestIp"))   { prefs.getString("nestIp",   nestIp,   sizeof(nestIp));   loaded++; }
   if (prefs.isKey("uploadToken")) { prefs.getString("uploadToken", uploadToken, sizeof(uploadToken)); loaded++; }
-  if (prefs.isKey("syncEvery"))           { syncEvery           = max(1, prefs.getInt("syncEvery", syncEvery)); loaded++; }
-  if (prefs.isKey("heartbeatIntervalMs")) { heartbeatIntervalMs = max(1000, prefs.getInt("heartbeatIntervalMs", heartbeatIntervalMs)); loaded++; }
+  if (prefs.isKey("syncEvery"))           { syncEvery           = max(1, (int)prefs.getInt("syncEvery", syncEvery)); loaded++; }
+  if (prefs.isKey("heartbeatIntervalMs")) { heartbeatIntervalMs = max(1000, (int)prefs.getInt("heartbeatIntervalMs", heartbeatIntervalMs)); loaded++; }
   if (prefs.isKey("verboseSerial"))       { verboseSerial = prefs.getBool("verboseSerial", verboseSerial); loaded++; }
 
   prefs.end();
